GlobalVarOffload: Moved GlobalVarOffload to data_offload.py
